+++ db/prefixdb.go
@@ -154,7 +154,7 @@ func (pdb *PrefixDB) prefixed(key []byte) []byte {
 
 // IteratePrefix is a convenience function for iterating over a key domain
 // restricted by prefix.
-func IteratePrefix(db DB, prefix []byte) (corestore.Iterator, error) {
+func IteratePrefix(db corestore.KVStoreWithBatch, prefix []byte) (corestore.Iterator, error) {
 	var start, end []byte
 	if len(prefix) == 0 {
 		start = nil