+++ immutable_tree.go
@@ -4,10 +4,7 @@ import (
 	"fmt"
 	"strings"
 
-	"cosmossdk.io/core/log"
 	corestore "cosmossdk.io/core/store"
-
-	dbm "github.com/cosmos/iavl/db"
 )
 
 // ImmutableTree contains the immutable tree at a given version. It is typically created by calling
@@ -17,7 +14,7 @@ import (
 // Returned key/value byte slices must not be modified, since they may point to data located inside
 // IAVL which would also be modified.
 type ImmutableTree struct {
-	logger log.Logger
+	logger Logger
 
 	root                   *Node
 	ndb                    *nodeDB
@@ -26,7 +23,7 @@ type ImmutableTree struct {
 }
 
 // NewImmutableTree creates both in-memory and persistent instances
-func NewImmutableTree(db dbm.DB, cacheSize int, skipFastStorageUpgrade bool, lg log.Logger, options ...Option) *ImmutableTree {
+func NewImmutableTree(db corestore.KVStoreWithBatch, cacheSize int, skipFastStorageUpgrade bool, lg Logger, options ...Option) *ImmutableTree {
 	opts := DefaultOptions()
 	for _, opt := range options {
 		opt(&opts)