+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Unreleased
+
+### Improvements
+
+- [#952](https://github.com/cosmos/iavl/pull/952) Add `DeleteVersionsFrom(int64)` API.
+- [#955](https://github.com/cosmos/iavl/pull/955) Get rid of `cosmos-db` deps completely.
+- [#961](https://github.com/cosmos/iavl/pull/961) Add new `GetLatestVersion` API to get the latest version.
+- [#965](https://github.com/cosmos/iavl/pull/965) Use expected interface for expected IAVL `Logger`.
+- [#970](https://github.com/cosmos/iavl/pull/970) Close the pruning process when the nodeDB is closed.
+- [#980](https://github.com/cosmos/iavl/pull/980) Use the `sdk/core/store.KVStoreWithBatch` interface instead of `iavl/db.DB` interface
+
 ## v1.2.0 May 13, 2024
 
 ### Improvements